Dumped the code I wrote.
There are implementations of several of the methods from the paper as
well as some of the recurrence relations. There's also a master
function I started writing to stitch things together, but it's *very*
incomplete.

The code is a bit janky (didn't originally intend to share it),
but hopefully it's readable. I figured it was best to share sooner
rather than later.